requirements.txt
gradio==4.31
websocket-client
pillow
requests
extcolors
matplotlib
numpy
pyperclip